work without github and google
