+++ src/authorisation/index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+    <title>Authorisation</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="./main.css">
+<body>
+<div class="body__auth-container">
+    <div id="firebaseui-auth-container"></div>
+</div>
+<img src="https://cdn-icons-png.flaticon.com/512/4193/4193251.png" >
+
+<div>
+    <script src="../../dist/auth.bundle.js"></script>
+    <script src="https://www.gstatic.com/firebasejs/ui/6.0.1/firebase-ui-auth.js"></script>
+    <link type="text/css" rel="stylesheet" href="https://www.gstatic.com/firebasejs/ui/6.0.1/firebase-ui-auth.css"/>
+</div>
+</body>
+
+</html>
+
