+++ src/connection/firebase.js
@@ -1,7 +1,7 @@
 import { initializeApp } from 'firebase/app'
 import { getDatabase, ref, set, push, onValue } from 'firebase/database'
 import getKey from '../util/utili18.js'
-import { ValidationFromLabels } from '../util/validation'
+import { ValidationFromLabels } from '../util/validation.js'
 
 export function initialisation (firebaseConfig) {
   initializeApp(firebaseConfig)
