+++ src/main/index.html
@@ -21,7 +21,6 @@
     <div class="content__canvas">
         <canvas class="content__canvas_graph" id="canvas" width="300" height="300"></canvas>
     </div>
-
     <form class="content__labels" >
         <div class="content__labels_x">
 
@@ -57,7 +56,6 @@
         </div>
         <button type="submit" class="content__labels_no-active" id="send" disabled="true">send</button>
     </form>
-    <button type="submit" class="content__labels_no-active" id="login">login</button>
     <div class="content__scroll">
         <div class="content__scroll_table" id="table">
             <div class="content__scroll_table_row" id="head_table">
@@ -71,12 +69,9 @@
         </div>
     </div>
 </div>
-<div id="firebaseui-auth-container"></div>
-<div>
-    <script src="../../dist/bundle.js"></script>
-    <script src="https://www.gstatic.com/firebasejs/ui/6.0.1/firebase-ui-auth.js"></script>
-    <link type="text/css" rel="stylesheet" href="https://www.gstatic.com/firebasejs/ui/6.0.1/firebase-ui-auth.css" />
 
+<div>
+    <script src="../../dist/main.bundle.js" ></script>
 </div>
 </body>
