--- src/main/indexx.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-    <script src="https://www.gstatic.com/firebasejs/9.1.3/firebase-app-compat.js"></script>
-    <script src="https://www.gstatic.com/firebasejs/9.1.3/firebase-auth-compat.js"></script>
-    <!-- *******************************************************************************************
-       * TODO(DEVELOPER): Paste the initialization snippet from this dialog box:
-       * Firebase Console > Project Settings > Add App > Web.
-       ***************************************************************************************** -->
-    <script src="https://www.gstatic.com/firebasejs/ui/6.0.1/firebase-ui-auth.js"></script>
-    <link type="text/css" rel="stylesheet" href="https://www.gstatic.com/firebasejs/ui/6.0.1/firebase-ui-auth.css" />
-    <script src="auth.js" type="mode"></script>
-</head>
-<body>
-<h1>Welcome to My Awesome App</h1>
-<div id="firebaseui-auth-container"></div>
-</body>
-</html>
